Change size_t to std::size_t

+++ kurdikova.elizaveta/I3/DeleteVowelLetters.cpp
@@ -3,7 +3,7 @@
 
 char * deleteVowels (char *result1, const char *source)
 {
-  int num = 0;
+  size_t num = 0;
   char vowelLetters[6] = {'a', 'e', 'i', 'o', 'u', 'y'};
   for (size_t i = 0; source[i] != '\0'; i++) {
     bool isVowel = false;
